f32 operators, reinterpret numeric bits
