+++ src/host/encode.rs
@@ -1,5 +1,9 @@
 use super::*;
-use crate::{ops::Op, run::Lab, util::maybe_grow};
+use crate::{
+  ops::{Num, Op},
+  run::Lab,
+  util::maybe_grow,
+};
 
 impl Host {
   /// Converts an ast net to a list of instructions to create the net.
@@ -99,16 +103,21 @@ impl<'a, E: Encoder> State<'a, E> {
         }
         self.encoder.link(l, r);
       }
-      Tree::Op { op, rhs: lft, out: rgt } => {
-        if let Tree::Int { val } = &**lft {
-          let o = self.encoder.op_int(*op, trg, *val);
+      Tree::Op { op, rhs: lft, out: rgt } => match &**lft {
+        Tree::Int { val } => {
+          let o = self.encoder.op_num(*op, trg, Num::Int(*val));
           self.visit_tree(rgt, o);
-        } else {
+        }
+        Tree::F32 { val } => {
+          let o = self.encoder.op_num(*op, trg, Num::Float(val.0));
+          self.visit_tree(rgt, o);
+        }
+        _ => {
           let (l, r) = self.encoder.op(*op, trg);
           self.visit_tree(lft, l);
           self.visit_tree(rgt, r);
         }
-      }
+      },
       Tree::Mat { zero, succ, out } => {
         let (a, o) = self.encoder.mat(trg);
         let (z, s) = self.encoder.ctr(0, a);
@@ -133,8 +142,7 @@ trait Encoder {
   fn make_const(&mut self, port: Port) -> Self::Trg;
   fn ctr(&mut self, lab: Lab, trg: Self::Trg) -> (Self::Trg, Self::Trg);
   fn op(&mut self, op: Op, trg: Self::Trg) -> (Self::Trg, Self::Trg);
-  fn op_int(&mut self, op: Op, trg: Self::Trg, rhs: i64) -> Self::Trg;
-  fn op_float(&mut self, op: Op, trg: Self::Trg, rhs: f32) -> Self::Trg;
+  fn op_num(&mut self, op: Op, trg: Self::Trg, rhs: Num) -> Self::Trg;
   fn mat(&mut self, trg: Self::Trg) -> (Self::Trg, Self::Trg);
   fn wires(&mut self) -> (Self::Trg, Self::Trg, Self::Trg, Self::Trg);
 }
@@ -172,14 +180,9 @@ impl Encoder for InterpretedDef {
     self.instr.push(Instruction::Op { op, trg, rhs, out });
     (rhs, out)
   }
-  fn op_int(&mut self, op: Op, trg: Self::Trg, rhs: i64) -> Self::Trg {
+  fn op_num(&mut self, op: Op, trg: Self::Trg, rhs: Num) -> Self::Trg {
     let out = self.new_trg_id();
-    self.instr.push(Instruction::OpInt { op, trg, rhs, out });
-    out
-  }
-  fn op_float(&mut self, op: Op, trg: Self::Trg, rhs: f32) -> Self::Trg {
-    let out = self.new_trg_id();
-    self.instr.push(Instruction::OpF32 { op, trg, rhs, out });
+    self.instr.push(Instruction::OpNum { op, trg, rhs, out });
     out
   }
   fn mat(&mut self, trg: Self::Trg) -> (Self::Trg, Self::Trg) {
@@ -215,11 +218,8 @@ impl<'a, M: Mode> Encoder for run::Net<'a, M> {
   fn op(&mut self, op: Op, trg: Self::Trg) -> (Self::Trg, Self::Trg) {
     self.do_op(op, trg)
   }
-  fn op_int(&mut self, op: Op, trg: Self::Trg, rhs: i64) -> Self::Trg {
-    self.do_op_int(op, trg, rhs)
-  }
-  fn op_float(&mut self, op: Op, trg: Self::Trg, rhs: f32) -> Self::Trg {
-    self.do_op_float(op, trg, rhs)
+  fn op_num(&mut self, op: Op, trg: Self::Trg, rhs: Num) -> Self::Trg {
+    self.do_op_num(op, trg, rhs)
   }
   fn mat(&mut self, trg: Self::Trg) -> (Self::Trg, Self::Trg) {
     self.do_mat(trg)
