+++ src/main.rs
@@ -467,7 +467,10 @@ fn compile_executable(target: &str, host: Arc<Mutex<host::Host>>) -> Result<(), 
     }
     lib
     main
-    ops
+    ops {
+      int
+      float
+    }
     run {
       addr
       allocator
